Restructured MDCTaskHandler to run consecutive Tests
