+++ hbp_with_firebase_mlkit/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:hbp_with_firebase_mlkit/CameraHandler.dart';
-import 'package:hbp_with_firebase_mlkit/TaskScreen.dart';
+import 'package:hbp_with_firebase_mlkit/MDCTaskHandler/TaskScreen.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/material.dart';
